requirements.txt
Flask==2.2.3
Flask_MySQLdb==1.0.1
paho_mqtt==1.6.1
Pillow==9.4.0
requests==2.28.2
Werkzeug==2.2.2
peewee==3.17.1
pywin32==306
pyinstaller==6.5.0